Refactor stdlib arithmetic into package
