+++ qamomile/circuit/stdlib/arithmetic/carry_venting.py
@@ -1,0 +1,223 @@
+"""Implement carry-venting constant addition primitives."""
+
+from __future__ import annotations
+
+import qamomile.circuit as qmc
+from qamomile.circuit.frontend.handle import Qubit, UInt, Vector
+from qamomile.circuit.frontend.handle.utils import get_size
+
+from .bitwise import (
+    _apply_offset_and,
+    _apply_offset_parity_and,
+    _apply_offset_parity_ccx,
+    _apply_offset_x,
+    _apply_two_offset_parities_ccx,
+    _phase_flip_if,
+)
+
+
+def _carry_xor_const(
+    target: Vector[Qubit],
+    dirty: Vector[Qubit],
+    value: UInt,
+    control: Qubit | None,
+) -> tuple[Qubit | None, Vector[Qubit], Vector[Qubit]]:
+    """XOR right-shifted addition carries into dirty workspace.
+
+    This is the carry-xor block from the constant-workspace classical-quantum
+    adder. ``dirty`` is restored by applying this block twice around the phase
+    corrections generated by carry venting.
+
+    Args:
+        target (Vector[Qubit]): Low bits of the post-addition target.
+        dirty (Vector[Qubit]): Dirty carry workspace to toggle.
+        value (UInt): Classical offset used by the addition.
+        control (Qubit | None): Optional coherent control for the offset.
+
+    Returns:
+        tuple[Qubit | None, Vector[Qubit], Vector[Qubit]]: Updated handles.
+    """
+    dirty_size = get_size(dirty)
+    for index in range(dirty_size - 1, 0, -1):
+        bit = (value // (2**index)) % 2
+        source = target[index]
+        previous = dirty[index - 1]
+        current = dirty[index]
+        control, previous, source, current = _apply_offset_parity_ccx(
+            control,
+            previous,
+            source,
+            current,
+            bit,
+        )
+        target[index] = source
+        dirty[index - 1] = previous
+        dirty[index] = current
+
+    for index in range(dirty_size):
+        bit = (value // (2**index)) % 2
+        current = dirty[index]
+        control, current = _apply_offset_x(control, current, bit)
+        dirty[index] = current
+
+    bit = value % 2
+    source = target[0]
+    current = dirty[0]
+    control, source, current = _apply_offset_parity_and(
+        control,
+        source,
+        current,
+        bit,
+    )
+    target[0] = source
+    dirty[0] = current
+
+    for index in range(1, dirty_size):
+        bit = (value // (2**index)) % 2
+        source = target[index]
+        previous = dirty[index - 1]
+        current = dirty[index]
+        control, source, previous, current = _apply_two_offset_parities_ccx(
+            control,
+            source,
+            previous,
+            current,
+            bit,
+        )
+        target[index] = source
+        dirty[index - 1] = previous
+        dirty[index] = current
+    return control, target, dirty
+
+
+def _dirty_const_add_extended(
+    target: Vector[Qubit],
+    overflow: Qubit,
+    dirty: Vector[Qubit],
+    clean_first: Qubit,
+    clean_second: Qubit,
+    value: UInt,
+    control: Qubit | None,
+) -> tuple[
+    Qubit | None,
+    Vector[Qubit],
+    Qubit,
+    Vector[Qubit],
+    Qubit,
+    Qubit,
+]:
+    """Add a classical value with two clean and linear dirty workspace.
+
+    Implements the carry-venting construction of Craig Gidney, "A
+    Classical-Quantum Adder with Constant Workspace and Linear Gates"
+    (arXiv:2507.23079), specialized to an ``overflow || target`` register.
+    X-basis carry measurements are corrected coherently, so the operation is
+    valid inside controlled modular arithmetic.
+
+    Args:
+        target (Vector[Qubit]): Little-endian low target bits.
+        overflow (Qubit): Most-significant target bit.
+        dirty (Vector[Qubit]): At least ``len(target) - 1`` dirty qubits,
+            restored exactly.
+        clean_first (Qubit): First clean streaming-carry qubit.
+        clean_second (Qubit): Second clean streaming-carry qubit.
+        value (UInt): Classical value to add modulo ``2**(len(target) + 1)``.
+        control (Qubit | None): Optional coherent control for the addition.
+
+    Returns:
+        tuple[Qubit | None, Vector[Qubit], Qubit, Vector[Qubit], Qubit,
+        Qubit]: Updated control and target with all workspace restored.
+
+    Raises:
+        ValueError: If the target is too small or dirty workspace is short.
+    """
+    target_size = get_size(target)
+    dirty_size = get_size(dirty)
+    if target_size < 2:
+        raise ValueError("carry-venting constant addition requires two low bits.")
+    if dirty_size < target_size - 1:
+        raise ValueError(
+            "carry-venting constant addition requires len(target) - 1 dirty qubits."
+        )
+    dirty_root = dirty
+    dirty = dirty_root[: target_size - 1]
+
+    for index in range(target_size):
+        bit = (value // (2**index)) % 2
+        target_bit = target[index]
+        control, target_bit = _apply_offset_x(control, target_bit, bit)
+        target[index] = target_bit
+    overflow_bit = (value // (2**target_size)) % 2
+    control, overflow = _apply_offset_x(control, overflow, overflow_bit)
+
+    vents: list[qmc.Bit] = []
+    for index in range(target_size):
+        bit = (value // (2**index)) % 2
+        source = target[index]
+        if index == 0:
+            next_carry = clean_second
+            control, source, next_carry = _apply_offset_and(
+                control,
+                source,
+                next_carry,
+                bit,
+            )
+            clean_second = next_carry
+        else:
+            current = clean_second if index % 2 == 1 else clean_first
+            next_is_overflow = index == target_size - 1
+            next_carry = (
+                overflow
+                if next_is_overflow
+                else (clean_first if index % 2 == 1 else clean_second)
+            )
+            control, source, current, next_carry = _apply_offset_parity_ccx(
+                control,
+                source,
+                current,
+                next_carry,
+                bit,
+            )
+            current, source = qmc.cx(current, source)
+            dirty_slot = dirty[index - 1]
+            current, dirty_slot = qmc.cx(current, dirty_slot)
+            dirty[index - 1] = dirty_slot
+            current = qmc.h(current)
+            current, vent = qmc.measure_reset(current)
+            vents.append(vent)
+            if index % 2 == 1:
+                clean_second = current
+            else:
+                clean_first = current
+            if next_is_overflow:
+                overflow = next_carry
+            elif index % 2 == 1:
+                clean_first = next_carry
+            else:
+                clean_second = next_carry
+        target[index] = source
+        if index < target_size - 1:
+            next_carry = clean_first if index % 2 == 1 else clean_second
+            control, next_carry = _apply_offset_x(control, next_carry, bit)
+            if index % 2 == 1:
+                clean_first = next_carry
+            else:
+                clean_second = next_carry
+        else:
+            control, overflow = _apply_offset_x(control, overflow, bit)
+
+    target = qmc.x(target)
+    overflow = qmc.x(overflow)
+    for index, vent in enumerate(vents):
+        dirty_bit = dirty[index]
+        dirty_bit = _phase_flip_if(vent, dirty_bit)
+        dirty[index] = dirty_bit
+    control, target, dirty = _carry_xor_const(target, dirty, value, control)
+    for index, vent in enumerate(vents):
+        dirty_bit = dirty[index]
+        dirty_bit = _phase_flip_if(vent, dirty_bit)
+        dirty[index] = dirty_bit
+    target = qmc.x(target)
+    overflow = qmc.x(overflow)
+    dirty_root[: target_size - 1] = dirty
+    return control, target, overflow, dirty_root, clean_first, clean_second
